fix KO
split strnstr tolower toupper

+++ ft_tolower.c
@@ -16,5 +16,5 @@ size_t	ft_tolower(int c)
 {
 	if (65 <= c && c <= 90)
 		c += 32;
-	return (c);
+	return ((size_t) c);
 }
