+++ ft_toupper.c
@@ -16,5 +16,5 @@ size_t	ft_toupper(int c)
 {
 	if (97 <= c && c <= 122)
 		c -= 32;
-	return (c);
+	return ((size_t) c);
 }
